# pkg/auth/gen/controllers/delete_session.gen.go
// Code generated by archesai. DO NOT EDIT.

package controllers

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/google/uuid"
	"github.com/oapi-codegen/runtime"

	"github.com/archesai/archesai/pkg/auth/gen/application"
	"github.com/archesai/archesai/pkg/server"
)

// ============================================================================
// DeleteSession - DELETE /auth/sessions/{id}
// ============================================================================

// DeleteSessionHandler is the HTTP handler for DeleteSession.
type DeleteSessionHandler struct {
	deleteSession application.DeleteSession
}

// NewDeleteSessionHandler creates a new HTTP handler.
func NewDeleteSessionHandler(deleteSession application.DeleteSession) *DeleteSessionHandler {
	return &DeleteSessionHandler{deleteSession: deleteSession}
}

// RegisterDeleteSessionRoute registers the HTTP route for DeleteSession.
func RegisterDeleteSessionRoute(mux *http.ServeMux, handler *DeleteSessionHandler) {
	mux.HandleFunc("DELETE /auth/sessions/{id}", handler.ServeHTTP)
}

// Request types

// Response types

type DeleteSessionResponse interface {
	VisitDeleteSessionResponse(w http.ResponseWriter) error
}

type DeleteSession204Response struct {
}

func (response DeleteSession204Response) VisitDeleteSessionResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type DeleteSession400Response struct {
	server.ProblemDetails
}

func (response DeleteSession400Response) VisitDeleteSessionResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type DeleteSession401Response struct {
	server.ProblemDetails
}

func (response DeleteSession401Response) VisitDeleteSessionResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(401)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type DeleteSession404Response struct {
	server.ProblemDetails
}

func (response DeleteSession404Response) VisitDeleteSessionResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(404)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type DeleteSession422Response struct {
	server.ProblemDetails
}

func (response DeleteSession422Response) VisitDeleteSessionResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(422)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type DeleteSession429Response struct {
	server.ProblemDetails
}

func (response DeleteSession429Response) VisitDeleteSessionResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(429)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type DeleteSession500Response struct {
	server.ProblemDetails
}

func (response DeleteSession500Response) VisitDeleteSessionResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

// ServeHTTP handles the DELETE /auth/sessions/{id} endpoint.
func (h *DeleteSessionHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	// Extract session ID from context for authenticated operations
	sessionID, ok := ctx.Value(server.SessionIDContextKey).(uuid.UUID)
	if !ok {
		errorResp := DeleteSession401Response{
			ProblemDetails: server.NewUnauthorizedResponse("session required", r.URL.Path),
		}
		if err := errorResp.VisitDeleteSessionResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Build input from request
	input := &application.DeleteSessionInput{}
	input.SessionID = sessionID

	// Path parameter "id"
	var id uuid.UUID
	if err := runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true}); err != nil {
		errorResp := DeleteSession400Response{
			ProblemDetails: server.NewBadRequestResponse(fmt.Sprintf("Invalid format for parameter id: %s", err), r.URL.Path),
		}
		if err := errorResp.VisitDeleteSessionResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.ID = id

	// Execute
	if err := h.deleteSession.Execute(ctx, input); err != nil {
		errorResp := DeleteSession500Response{
			ProblemDetails: server.NewInternalServerErrorResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitDeleteSessionResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	response := DeleteSession204Response{}
	if err := response.VisitDeleteSessionResponse(w); err != nil {
		fmt.Fprintf(w, "error writing response: %v", err)
	}
}