# pkg/auth/gen/controllers/delete_user.gen.go
// Code generated by archesai. DO NOT EDIT.

package controllers

import (
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/google/uuid"
	"github.com/oapi-codegen/runtime"

	"github.com/archesai/archesai/pkg/auth/gen/application"
	"github.com/archesai/archesai/pkg/server"
)

// ============================================================================
// DeleteUser - DELETE /users/{id}
// ============================================================================

// DeleteUserHandler is the HTTP handler for DeleteUser.
type DeleteUserHandler struct {
	deleteUser application.DeleteUser
}

// NewDeleteUserHandler creates a new HTTP handler.
func NewDeleteUserHandler(deleteUser application.DeleteUser) *DeleteUserHandler {
	return &DeleteUserHandler{deleteUser: deleteUser}
}

// RegisterDeleteUserRoute registers the HTTP route for DeleteUser.
func RegisterDeleteUserRoute(mux *http.ServeMux, handler *DeleteUserHandler) {
	mux.HandleFunc("DELETE /users/{id}", handler.ServeHTTP)
}

// Request types

// Response types

type DeleteUserResponse interface {
	VisitDeleteUserResponse(w http.ResponseWriter) error
}

type DeleteUser204Response struct {
}

func (response DeleteUser204Response) VisitDeleteUserResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type DeleteUser400Response struct {
	server.ProblemDetails
}

func (response DeleteUser400Response) VisitDeleteUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type DeleteUser401Response struct {
	server.ProblemDetails
}

func (response DeleteUser401Response) VisitDeleteUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(401)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type DeleteUser404Response struct {
	server.ProblemDetails
}

func (response DeleteUser404Response) VisitDeleteUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(404)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type DeleteUser422Response struct {
	server.ProblemDetails
}

func (response DeleteUser422Response) VisitDeleteUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(422)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type DeleteUser429Response struct {
	server.ProblemDetails
}

func (response DeleteUser429Response) VisitDeleteUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(429)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

type DeleteUser500Response struct {
	server.ProblemDetails
}

func (response DeleteUser500Response) VisitDeleteUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	return json.NewEncoder(w).Encode(response.ProblemDetails)
}

// ServeHTTP handles the DELETE /users/{id} endpoint.
func (h *DeleteUserHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	// Extract session ID from context for authenticated operations
	sessionID, ok := ctx.Value(server.SessionIDContextKey).(uuid.UUID)
	if !ok {
		errorResp := DeleteUser401Response{
			ProblemDetails: server.NewUnauthorizedResponse("session required", r.URL.Path),
		}
		if err := errorResp.VisitDeleteUserResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	// Build input from request
	input := &application.DeleteUserInput{}
	input.SessionID = sessionID

	// Path parameter "id"
	var id uuid.UUID
	if err := runtime.BindStyledParameterWithOptions("simple", "id", r.PathValue("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true}); err != nil {
		errorResp := DeleteUser400Response{
			ProblemDetails: server.NewBadRequestResponse(fmt.Sprintf("Invalid format for parameter id: %s", err), r.URL.Path),
		}
		if err := errorResp.VisitDeleteUserResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}
	input.ID = id

	// Execute
	if err := h.deleteUser.Execute(ctx, input); err != nil {
		errorResp := DeleteUser500Response{
			ProblemDetails: server.NewInternalServerErrorResponse(err.Error(), r.URL.Path),
		}
		if err := errorResp.VisitDeleteUserResponse(w); err != nil {
			fmt.Fprintf(w, "error writing response: %v", err)
		}
		return
	}

	response := DeleteUser204Response{}
	if err := response.VisitDeleteUserResponse(w); err != nil {
		fmt.Fprintf(w, "error writing response: %v", err)
	}
}